chore: changing the position of callout
